requirements.txt
mcp[cli]
pymupdf
pytesseract
playwright
ultralytics
chromadb
pillow
pydantic
watchdog
